# BACKEND/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "cuidarteplus API (español)",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:4444" }],
  "paths": {
...
    "/documentos": {
      "post": {
        "summary": "Subir documento de examen",
        "description": "Sube un documento asociado a un examen médico. Requiere rol de médico o admin. Solo se permite un archivo por examen (PDF, JPG, PNG).",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "examen_medico_id": { "type": "integer" },
                  "paciente_id": { "type": "integer" },
                  "documento": {
                    "type": "string",
                    "format": "binary",
                    "description": "Un archivo (PDF, JPG, PNG)"
                  }
                },
                "required": ["examen_medico_id", "paciente_id", "documento"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Documento subido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documento": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "nombre_archivo": { "type": "string" },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Error de validación" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" }
        }
      }
    },
    "/documentos/examen/{examen_id}": {
      "get": {
        "summary": "Listar documentos de un examen",
        "description": "Lista todos los documentos asociados a un examen médico. Médicos y admins pueden ver documentos de cualquier examen. Pacientes solo pueden ver documentos de sus propios exámenes.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "examen_id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" }
        }
      }
    },
    "/documentos/{id}": {
      "get": {
        "summary": "Descargar documento",
        "description": "Descarga un documento específico. Médicos y admins pueden descargar cualquier documento. Pacientes solo pueden descargar sus propios documentos.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Archivo (PDF, JPG o PNG)",
            "content": {
              "application/pdf": {
                "schema": { "type": "string", "format": "binary" }
              },
              "image/jpeg": {
                "schema": { "type": "string", "format": "binary" }
              },
              "image/png": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" },
          "404": { "description": "Documento no encontrado" }
        }
      },
      "delete": {
        "summary": "Eliminar documento",
        "description": "Elimina un documento. Solo admin puede eliminar documentos.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Documento eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" },
          "404": { "description": "Documento no encontrado" }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido del endpoint /autenticacion/login. Incluye: usuarioId, nombreUsuario, rolId, rolNombre, pacienteId, roles"
      }
    }
  }
}